pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sage"
version = "0.1.0"
description = "Self-improving Autonomous Generation Engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "ollama>=0.3",
    "anthropic>=0.30",
    "openai>=1.30",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "qdrant-client>=1.9",        # Phase 3+
    "rich>=13.0",
    "tree-sitter>=0.25.0",       # Phase 4: semantic code reader
    "tree-sitter-languages>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "scikit-learn>=1.3",
]
rl = [
    "scikit-learn>=1.3",
]

[project.scripts]
sage = "sage.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sage = ["assets/skills/**/*.md"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["src/sage/assets/skills"]

[tool.mypy]
python_version = "3.11"
strict = false
